+++ CHANGELOG.md
@@ -8,6 +8,7 @@ a concise history of changes made.
 ## Latest
 - Added the 'enabled' field to Observer so loading Checklists can be blocked.
 - Added caching to the Latest, Weekly and Monthly News pages.
+- Added date range fields to the Checklists search panel.
 
 ## 2025.05.21
 - Added site logo.